refactor about page
