+++ src/api/index.js
@@ -1,6 +1,9 @@
 const axios = require('axios')
 
-export const test = async ({ objkt_id }) => {
+/**
+ * Get OBJKT detail page
+ */
+export const GetOBJKT = async ({ objkt_id }) => {
   return new Promise((resolve, reject) => {
     axios
       .post(process.env.REACT_APP_OBJKT, {
