+++ src/components/layout/container/index.module.scss
@@ -1,15 +1,13 @@
-@import '../../styles/variables.scss';
-@import '../../styles/layout.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/layout.scss';
 
 .container {
   width: 100%;
+  margin: 0 auto;
   margin-bottom: 20px;
-  border: 1px dashed white;
 
   @include respond-to('tablet') {
     width: 50%;
-    margin: 0 auto;
-    margin-bottom: 20px;
   }
 
   &.large {
