+++ src/components/layout/index.js
@@ -1,0 +1,3 @@
+export { Page } from './page'
+export { Container } from './container'
+export { Padding } from './padding'
