+++ src/components/layout/page/index.module.scss
@@ -1,0 +1,8 @@
+@import '../../../styles/variables.scss';
+@import '../../../styles/layout.scss';
+
+.container {
+  width: 100%;
+  min-height: 100vh;
+  padding: 100px 0;
+}
