+++ src/pages/about/index.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react'
-import { Container } from '../../components/container'
-import { Padding } from '../../components/padding'
+import { Page, Container, Padding } from '../../components/layout'
+import { Button } from '../../components/button'
 import { HicetnuncContext } from '../../context/HicetnuncContext'
+import styles from './index.module.scss'
 
 export default class About extends Component {
   static contextType = HicetnuncContext
@@ -18,75 +19,59 @@ export default class About extends Component {
 
   render() {
     return (
-      <Container>
-        <Padding>Hello world</Padding>
-      </Container>
+      <Page>
+        <Container>
+          <Padding>
+            <Button href="https://github.com/hicetnunc2000">
+              <strong>hic et nunc stack</strong>
+            </Button>
+          </Padding>
+        </Container>
+
+        <Container>
+          <Padding>
+            <p>
+              the present decentralized application allows its users to manage
+              decentralized digital assets, serving as a public smart contract
+              infrastructure on Tezos Blockchain.
+            </p>{' '}
+          </Padding>
+        </Container>
+
+        <Container>
+          <Padding>
+            <p>
+              IPFS NFTs can be minted and traded by permissionless means. such
+              experiment was designed intending to imagine alternative crypto
+              economies.
+            </p>{' '}
+          </Padding>
+        </Container>
+
+        <Container>
+          <Padding>
+            <p>
+              we're concerned about your security and autonomy. please verify
+              informations while making transactions.
+            </p>
+          </Padding>
+        </Container>
+
+        <Container>
+          <Padding>
+            <p>for consulting, networking or questions:</p>
+          </Padding>
+        </Container>
+
+        <Container>
+          <Padding>
+            <div className={styles.buttons}>
+              <Button href="https://discord.gg/jKNy6PynPK">discord</Button>
+              <Button href="mailto:hicetnunc2000@protonmail.com">email</Button>
+            </div>
+          </Padding>
+        </Container>
+      </Page>
     )
   }
-
-  // render() {
-  //   return (
-  //     <Row>
-  //       <Col sm="12" md={{ size: 6, offset: 3 }}>
-  //         <div style={{ animation: 'fadeMe 1.2s' }}>
-  //           <div style={{ padding: '15% 0', border: 0, textAlign: 'justify' }}>
-  //             <p style={{ fontWeight: 'bold' }}></p>
-  //             <p style={{ fontWeight: 'bold' }}>
-  //               <a
-  //                 style={{
-  //                   color: '#000',
-  //                   '&:hover': {
-  //                     color: '#000',
-  //                   },
-  //                 }}
-  //                 href="https://github.com/hicetnunc2000"
-  //               >
-  //                 <i>hic et nunc</i> stack
-  //               </a>
-  //             </p>
-  //             <p>
-  //               the present decentralized application allows its users to manage
-  //               decentralized digital assets, serving as a public smart contract
-  //               infrastructure on Tezos Blockchain.
-  //             </p>
-  //             <p>
-  //               IPFS NFTs can be minted and traded by permissionless means. such
-  //               experiment was designed intending to imagine alternative crypto
-  //               economies.
-  //             </p>
-  //             <p>
-  //               we're concerned about your security and autonomy. please verify
-  //               informations while making transactions.
-  //             </p>
-  //             <p>for consulting, networking or questions:</p>
-  //             <a
-  //               style={{
-  //                 color: '#000',
-  //                 '&:hover': {
-  //                   color: '#000',
-  //                 },
-  //               }}
-  //               href="https://discord.gg/jKNy6PynPK"
-  //             >
-  //               discord
-  //             </a>
-  //             <br />
-  //             <a
-  //               style={{
-  //                 color: '#000',
-  //                 '&:hover': {
-  //                   color: '#000',
-  //                 },
-  //               }}
-  //               href="mailto:hicetnunc2000@protonmail.com"
-  //             >
-  //               hicetnunc2000@protonmail.com
-  //             </a>
-  //             <br />
-  //           </div>
-  //         </div>
-  //       </Col>
-  //     </Row>
-  //   )
-  // }
 }
