+++ src/pages/about/index.module.scss
@@ -1,5 +1,4 @@
-@import '../../styles/layout.scss';
-@import '../../styles/variables.scss';
-
-.container {
+.buttons {
+  display: flex;
+  flex-direction: column;
 }
